converter/requirements.txt
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
idna==3.7
mysql-connector-python==8.4.0
pdfminer.six==20231228
pdfplumber==0.11.1
pillow==10.3.0
pycparser==2.22
pypdfium2==4.30.0
requests==2.32.3
urllib3==2.2.2
